refactor: drop the unreachable kvStorage Get/Set/SetNX seam methods
Neither Get nor Set had a production caller: they were exercised only by
tests written against themselves. SetNX had none either, because
MigrateV1ToV2 takes its lock through redisClient.SetNX directly rather
than through the storage seam, so the interface copy was never on a code
path. Removing all three also removes three methods from countingStorage,
the test fake that has to mirror the interface by hand.

The tests that used Set/Get as fixture setup now write through the
go-redis client already in scope, so what they assert - Del and Exists -
is unchanged. TestRedisStorageSetNX tested only the removed method and
goes with it.

Also moves stringRuneSource beside its only caller in the engine test
file, and swaps the deprecated sort.Ints/sort.Strings/sort.Slice for
slices.Sort/slices.SortFunc. api/v1.txt and NOTICE regenerate
byte-identical, so neither diff gate moves.

kvStorage is unexported and was withdrawn from the public surface before
v1.5.0 froze it, so the public API snapshot is untouched.
